fix(rpc): extract shared unary dispatch path
dispatcher.ts crossed the 300-line lint cap, breaking static analysis on
main and every PR branched from it. The unary path was duplicated between
dispatch() and the non-streaming branch of dispatchStreaming(); extract it
into dispatcher-unary-execution.ts and call it from both.
